Add item caching
Serialises the inventory/item pairs to a table when the inventory
changes and loads it again on start. This requires the cacheItems config
option to be enabled.

Peripheral loading is deferred until the main thread has started as a high
priority task.

Also use Artist's serialisation library in a couple of other places as it
produces smaller output.