Rewrite the UI system
Refactor the generic UI system (layers, drawing, etc...) out of
interface.lua into separate non-artist-specific framework (ignoring our
class library).

The UI system works on a series of "layers" - effectively a stack of
screens. The top layer is the active one - it is rendered in bright
colours and is the only layer to receive events. Other layers below are
rendered dimmed out. Typically we only have one or two layers
(the main UI and a dialogue), but this could technically support any
number.

Components in the UI are standard classes, though this is more for
convenience rather than any adherence to OOP. There's no interfaces or
inheritance (boo OOP, but mostly I'm too lazy) - instead we allow
methods to be optional, only calling them when the exist.

Components can be focused or blurred if they provide a `:focus`
method. We find all focusable components when adding a layer, and allow
tab/Shift+Tab to switch between them.

If a component is focused, we look at its (and its parents') `keymap`
field and merge them together (using metis's keybinding module). This
provides a friendlier way of handling keybindings than hard-coding it
like we did in the old read.lua module.

We also define a couple of built-in components (Button, Frame, Input),
which allows us to build something like dialogue.lua in a much more
simple and reusable way.